pyproject.toml
[project]
name = "uart-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.12.2",
    "pyqt6>=6.9.1",
    "pyserial>=3.5",
]
